RAG/requirements.txt
langchain
fastapi
langchain-chroma
langchain-community
langchain-ollama
langchain-google-genai
python-dotenv